[sdk] Rename createVoucher -> createSignableVoucher, Move to resolve-signatures
