+++ packages/sdk/src/resolve/resolve-signatures.js
@@ -1,6 +1,5 @@
 import {isTransaction} from "../interaction/interaction.js"
-import {createVoucher} from "../utils"
-import {sansPrefix} from "@onflow/util-address"
+import {sansPrefix, withPrefix} from "@onflow/util-address"
 import {
   encodeTransactionPayload as encodeInsideMessage,
   encodeTransactionEnvelope as encodeOutsideMessage,
@@ -59,6 +58,31 @@ function fetchSignature(ix, payload) {
   }
 }
 
+export const createSignableVoucher = ix => {
+  return {
+    cadence: ix.message.cadence,
+    refBlock: ix.message.refBlock || null,
+    computeLimit: ix.message.computeLimit,
+    arguments: ix.message.arguments.map(id => ix.arguments[id].asArgument),
+    proposalKey: {
+      address: withPrefix(ix.accounts[ix.proposer].addr),
+      keyId: ix.accounts[ix.proposer].keyId,
+      sequenceNum: ix.accounts[ix.proposer].sequenceNum,
+    },
+    payer: withPrefix(ix.accounts[ix.payer].addr),
+    authorizers: ix.authorizations
+      .map(cid => withPrefix(ix.accounts[cid].addr))
+      .reduce((prev, current) => {
+        return prev.find(item => item === current) ? prev : [...prev, current]
+      }, []),
+    payloadSigs: findInsideSigners(ix).map(id => ({
+      address: withPrefix(ix.accounts[id].addr),
+      keyId: ix.accounts[id].keyId,
+      sig: ix.accounts[id].signature,
+    })),
+  }
+}
+
 export function buildSignable(acct, message, ix) {
   try {
     return {
@@ -72,7 +96,7 @@ export function buildSignable(acct, message, ix) {
       args: ix.message.arguments.map(d => ix.arguments[d].asArgument),
       data: {},
       interaction: ix,
-      voucher: createVoucher(ix),
+      voucher: createSignableVoucher(ix),
     }
   } catch (error) {
     console.error("buildSignable", error)
