--- packages/sdk/src/utils/create-voucher.js
@@ -1,26 +1,0 @@
-import {withPrefix} from "@onflow/util-address"
-import {findInsideSigners} from "../resolve/resolve-signatures"
-export const createVoucher = ix => {
-  return {
-    cadence: ix.message.cadence,
-    refBlock: ix.message.refBlock || null,
-    computeLimit: ix.message.computeLimit,
-    arguments: ix.message.arguments.map(id => ix.arguments[id].asArgument),
-    proposalKey: {
-      address: withPrefix(ix.accounts[ix.proposer].addr),
-      keyId: ix.accounts[ix.proposer].keyId,
-      sequenceNum: ix.accounts[ix.proposer].sequenceNum,
-    },
-    payer: withPrefix(ix.accounts[ix.payer].addr),
-    authorizers: ix.authorizations
-      .map(cid => withPrefix(ix.accounts[cid].addr))
-      .reduce((prev, current) => {
-        return prev.find(item => item === current) ? prev : [...prev, current]
-      }, []),
-    payloadSigs: findInsideSigners(ix).map(id => ({
-      address: withPrefix(ix.accounts[id].addr),
-      keyId: ix.accounts[id].keyId,
-      sig: ix.accounts[id].signature,
-    })),
-  }
-}
